feat: refactor onboarding forms and navigation
