+++ src/app/onboarding/[step]/page.tsx
@@ -1,4 +1,5 @@
-import { notFound, redirect } from 'next/navigation';
+import { Metadata } from 'next';
+import { notFound } from 'next/navigation';
 import React from 'react';
 
 import { Step1PathDecision } from '@/components/onboarding/Step1PathDecision';
@@ -7,13 +8,16 @@ import { Step2PoliticalRoleDetails } from '@/components/onboarding/Step2Politica
 import { Step3FocusArea } from '@/components/onboarding/Step3FocusArea';
 import { Step4Completion } from '@/components/onboarding/Step4Completion';
 import { Step5Preview } from '@/components/onboarding/Step5Preview';
+import { Progress } from '@/components/ui/progress';
+import { getCurrentProfile } from '@/domain/actions/onboarding';
+export const metadata: Metadata = {
+  title: 'Onboarding - OpenEU',
+  description: 'Complete your OpenEU profile setup',
+};
 
 interface OnboardingStepPageProps {
   params: Promise<{
     step: string;
-  }>;
-  searchParams: Promise<{
-    userCategory?: 'entrepreneur' | 'politician';
   }>;
 }
 
@@ -22,10 +26,8 @@ const TOTAL_STEPS = 5;
 
 export default async function OnboardingStepPage({
   params,
-  searchParams,
 }: OnboardingStepPageProps) {
   const { step } = await params;
-  const { userCategory } = await searchParams;
 
   // Validate step parameter
   if (!VALID_STEPS.includes(step)) {
@@ -34,25 +36,23 @@ export default async function OnboardingStepPage({
 
   const currentStep = parseInt(step, 10);
 
-  // For step 2 and beyond, userCategory is required
-  if (currentStep > 1 && !userCategory) {
-    redirect('/onboarding/1');
-  }
-
   // Render the appropriate step component
-  const renderStep = () => {
+  const renderStep = async () => {
     switch (currentStep) {
       case 1:
         return <Step1PathDecision />;
-      case 2:
-        if (userCategory === 'entrepreneur') {
-          return <Step2EntrepreneurRoleDetails />;
-        } else if (userCategory === 'politician') {
+      case 2: {
+        // Get user's profile to determine which Step2 component to show
+        const profileResult = await getCurrentProfile();
+        const userCategory = profileResult.data?.userCategory;
+
+        if (userCategory === 'politician') {
           return <Step2PoliticalRoleDetails />;
         } else {
-          redirect('/onboarding/1');
+          // Default to entrepreneur if no category set or if entrepreneur
+          return <Step2EntrepreneurRoleDetails />;
         }
-        break;
+      }
       case 3:
         return <Step3FocusArea />;
       case 4:
@@ -77,16 +77,11 @@ export default async function OnboardingStepPage({
               {Math.round((currentStep / TOTAL_STEPS) * 100)}% Complete
             </span>
           </div>
-          <div className="w-full bg-muted rounded-full h-2">
-            <div
-              className="bg-primary h-2 rounded-full transition-all duration-300"
-              style={{ width: `${(currentStep / TOTAL_STEPS) * 100}%` }}
-            />
-          </div>
+          <Progress value={(currentStep / TOTAL_STEPS) * 100} className="h-2" />
         </div>
 
         {/* Step content */}
-        <div>{renderStep()}</div>
+        <div>{await renderStep()}</div>
       </div>
     </div>
   );
