--- src/app/onboarding/page.tsx
@@ -1,6 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default async function OnboardingPage() {
-  // Redirect to the first step of the new onboarding flow
-  redirect('/onboarding/1');
-}
